+++ extensions/notion/src/components/actions/index.ts
@@ -1,3 +1,3 @@
-export * from "./ActionEditPageProperty";
+export * from "./ActionKanbanStatus";
 export * from "./ActionSetVisibleProperties";
 export * from "./ActionSetOrderProperties";